# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "iApp API Documentation",
        "description": "API Documentation - iApp",
        "contact": {
            "email": "akash.corp@akij.net"
        },
        "license": {
            "name": "@iapp.akij.net",
            "url": "https://iapp.akij.net"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/business/{id}": {
            "get": {
                "tags": [
                    "Business"
                ],
                "summary": "Get Business By ID",
                "description": "Get Business By ID",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Business By ID"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Business"
                ],
                "summary": "Update Business",
                "description": "Update Business",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "bin": {
                                        "type": "string"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "example": "2020-01-01"
                                    },
                                    "tax_label_1": {
                                        "type": "string",
                                        "example": "Tax"
                                    },
                                    "tax_label_2": {
                                        "type": "string",
                                        "example": "Tax"
                                    },
                                    "tax_number_1": {
                                        "type": "string",
                                        "example": "Tax"
                                    },
                                    "tax_number_2": {
                                        "type": "string"
                                    },
                                    "default_profit_percent": {
                                        "type": "integer"
                                    },
                                    "owner_id": {
                                        "type": "string",
                                        "example": 1
                                    },
                                    "time_zone": {
                                        "type": "string"
                                    },
                                    "fy_start_month": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "default_sales_discount": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "accounting_method": {
                                        "type": "string",
                                        "example": "FIFO"
                                    },
                                    "sell_price_tax": {
                                        "type": "string",
                                        "example": "includes"
                                    },
                                    "currency_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "logo": {
                                        "type": "string"
                                    },
                                    "sku_prefix": {
                                        "type": "string"
                                    },
                                    "enable_tooltip": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Business"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Business"
                ],
                "summary": "Delete Business",
                "description": "Delete Business",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Business"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/api/v1/locations/{id}": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Get Location By ID",
                "description": "Get Location By ID",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Location By ID"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Locations"
                ],
                "summary": "Update Location",
                "description": "Update Location",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "business_id": {
                                        "type": "integer"
                                    },
                                    "country": {
                                        "type": "string"
                                    },
                                    "state": {
                                        "type": "string"
                                    },
                                    "zip_code": {
                                        "type": "string"
                                    },
                                    "landmark": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    },
                                    "alternate_number": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Location"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Locations"
                ],
                "summary": "Delete Location",
                "description": "Delete Location",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Location"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/api/v1/currencies/{id}": {
            "get": {
                "tags": [
                    "Currency"
                ],
                "summary": "Get Currency By ID",
                "description": "Get Currency By ID",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Currency By ID"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Currency"
                ],
                "summary": "Update Currency",
                "description": "Update Currency",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "country": {
                                        "type": "string"
                                    },
                                    "currency": {
                                        "type": "string"
                                    },
                                    "code": {
                                        "type": "string"
                                    },
                                    "symbol": {
                                        "type": "string"
                                    },
                                    "thousand_separator": {
                                        "type": "string"
                                    },
                                    "decimal_separator": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Currency"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Currency"
                ],
                "summary": "Delete Currency",
                "description": "Delete Currency",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Currency"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/api/v1/tax/{id}": {
            "get": {
                "tags": [
                    "Tax"
                ],
                "summary": "Get Tax By ID",
                "description": "Get Tax By ID",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Tax By ID"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tax"
                ],
                "summary": "Update Tax",
                "description": "Update Tax",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tax"
                                    },
                                    "business_id": {
                                        "type": "integer"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "example": "2020-01-01"
                                    },
                                    "calculation_type": {
                                        "type": "string",
                                        "example": "Tax"
                                    },
                                    "amount": {
                                        "type": "string",
                                        "example": "10.00"
                                    },
                                    "is_tax_group": {
                                        "type": "boolean"
                                    },
                                    "rounding_type": {
                                        "type": "string"
                                    },
                                    "created_by": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Tax"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tax"
                ],
                "summary": "Delete Tax",
                "description": "Delete Tax",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Tax"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/api/v1/attributes/{id}": {
            "get": {
                "tags": [
                    "Attribute"
                ],
                "summary": "Get Attribute By ID",
                "description": "Get Attribute By ID",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Attribute By ID"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Attributes"
                ],
                "summary": "Update Attribute",
                "description": "Update Attribute",
                "operationId": "Modules\\Item\\Http\\Controllers\\AttributeController::update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "business_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Attribute"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Attributes"
                ],
                "summary": "Delete Attribute",
                "description": "Delete Attribute",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Attribute"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/api/v1/brands/business/{business_id}": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Get Brand List of Business",
                "description": "Get Brand List of Business",
                "operationId": "getBrandByBusiness",
                "parameters": [
                    {
                        "name": "business_id",
                        "in": "path",
                        "description": "business_id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Brand List of Business"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            },
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}